chore: Init data.rs
